add .env
changed id type from Long to int

added cors config

+++ src/main/java/com/mahan/springbackend/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.mahan.springbackend.exception;
 
 public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException(Long id) {
+    public UserNotFoundException(int id) {
         super("Can't find user with id " + id);
     }
 }
